Removed dependency on gzip files in gs_patterns_core.cpp, moved that to gs_patterns.cpp which handles trace file based input. Refatored second_pass handling so it can be done with any code able to get a collection of trace_t's.
