+++ gs_patterns_core.h
@@ -38,4 +38,7 @@ int get_top_target(InstrInfo & target_iinfo, Metrics & target_metrics);
 
 void normalize_stats(Metrics & target_metrics);
 
-void second_pass(gzFile fp_drtrace, Metrics & gather_metrics, Metrics & scatter_metrics);
+bool handle_2nd_pass_trace_entry(trace_entry_t * drline,
+                                 Metrics & gather_metrics, Metrics & scatter_metrics,
+                                 addr_t & iaddr, int64_t & maddr, uint64_t & mcnt,
+                                 addr_t * gather_base, addr_t * scatter_base);
